requirements.txt
jsonlines>=1.2.0,<2.0.0
lxml>=4.5.0,<5.0.0
nltk>=3.4.5,<4.0.0
progressbar2>=3.51.4,<4.0.0
protego>=0.1.16,<0.2.0
requests>=2.23.0,<3.0.0
requests-cache>=0.9.3,<1.0.0
typeguard>=2.7.1,<3.0.0
